Use ChannelFactory
 - ChannelFactory to create message channels and configure message converters
