+++ spring-cloud-stream/src/main/java/org/springframework/cloud/stream/config/ChannelBindingServiceConfiguration.java
@@ -31,6 +31,7 @@ import org.springframework.cloud.stream.binder.Binder;
 import org.springframework.cloud.stream.binding.BinderAwareChannelResolver;
 import org.springframework.cloud.stream.binding.BinderAwareRouterBeanPostProcessor;
 import org.springframework.cloud.stream.binding.ChannelBindingService;
+import org.springframework.cloud.stream.binding.ChannelFactory;
 import org.springframework.cloud.stream.binding.ContextStartAfterRefreshListener;
 import org.springframework.cloud.stream.binding.InputBindingLifecycle;
 import org.springframework.cloud.stream.binding.OutputBindingLifecycle;
@@ -68,6 +69,11 @@ public class ChannelBindingServiceConfiguration {
 			ChannelBindingServiceProperties channelBindingServiceProperties,
 			Binder<MessageChannel> binder) {
 		return new ChannelBindingService(channelBindingServiceProperties, binder);
+	}
+
+	@Bean
+	public ChannelFactory channelFactory(ChannelBindingServiceProperties channelBindingServiceProperties) {
+		return new ChannelFactory(channelBindingServiceProperties);
 	}
 
 	@Bean
